libticalcs: add dusb_recv_data_varsize function, and torture tests.
This is a variant of dusb_recv_data that allows packets to be larger
or smaller than the size declared in the packet header.  (In most
cases this is an error, but 84+CSE screenshot packets are an
exception, because the total size can't be predicted ahead of time.)

In addition, rewrite dusb_recv_data to use dusb_recv_data_varsize,
treating it as an error if the declared size does not match the actual
size.

Code by Benjamin, torture tests by Lionel.

Signed-off-by: Benjamin Moody <floppusmaximus@users.sf.net>
Signed-off-by: Lionel Debroux <lionel_debroux@yahoo.fr>