libticalcs: add support for 84+CSE screenshots.
For both DBUS and DUSB, the type of screenshot is determined based on
the size of the packet the calc sends back.  In the DBUS case, note
that the calc will send many packets, each of which (at least in the
current implementation) will be no larger than 250 bytes, and the end
of the sequence is marked with an EOT.  (The traditional protocol, in
contrast, uses a single 768-byte packet and no EOT.  This means we
have to guess, after receiving the first packet, whether to expect
more packets or not.)

The DUSB case uses dusb_cmd_r_screenshot, which receives the entire
screenshot as a single (non-standard) packet.

In addition, a 'pixel_format' field has been added to the
CalcScreenCoord struct, defining the format of pixel data (currently
including monochrome, 4-bit grayscale, or 16-bit RGB.)  All versions
of recv_screen have been updated to set this field appropriately.
(NOTE: this change constitutes an ABI break!)

Code by Benjamin, French translation by Lionel.

Signed-off-by: Benjamin Moody <floppusmaximus@users.sf.net>
Signed-off-by: Lionel Debroux <lionel_debroux@yahoo.fr>