libtifiles: handle version compatibility byte for 83+/84+ variables.
This byte is set for each variable to indicate the required OS version
(e.g., based on which tokens are used in a program.)  Most third-party
programs simply set this byte to zero, which is usually harmless;
however, 84+CSE Pic variables must be set to version 10, and there may
be similar issues with other data types in the future.

When reading an 83+/84+ file, automatically promote "version 0" Pic
variables (created by older versions of libtifiles) to version 10 if
they are the correct size for an 84+CSE picture.