libticalcs: define attribute 0x13 as AID_ARCHIVED2; omit harmful 'archived' attribute when deleting variables on DUSB calculators (84+ family, 83PCE/84+CE, 89T).
The 0x13 attribute is to AID_ARCHIVED as AID_VAR_TYPE2 is to AID_VAR_TYPE; to be specific, it can be used in DUSB_VPKT_VAR_REQ and DUSB_VPKT_DEL_VAR commands to specify whether the target variable is currently archived.
This attribute is in no way required when deleting variables from an 84+, 84+CSE or 89T. On the 84+CE, this attribute seems not to work at all.
Moreover, it's not usually desirable to restrict deletion based on whether the variable is archived: for example, it makes tirm fail when given an explicit variable name.
Removing the attribute brings DUSB into line with the way DBUS remote deletion has always worked.

Lionel: squashed Benjamin's patches, tested on 89T.

Signed-off-by: Benjamin Moody <floppusmaximus@users.sf.net>
Signed-off-by: Lionel Debroux <lionel_debroux@yahoo.fr>