libtifiles: better handling of non-ASCII application names for 84+ family.
To be consistent with the handling of variable names in FileContent,
expand the name field of FlashContent to allow UTF-8 names, and add a
model_dst field that, as in FileContent, indicates which encoding is
being used.

Please note that this is a major ABI break and most applications will
need to be recompiled.

When reading an app file, attempt to parse the name from the internal
app header (which is the name that the calculator actually uses to
refer to the app.)  Use this name in preference to the name indicated
by the TIFL header, since the latter may be unreliable.

When writing an app file, attempt to parse the name from the internal
header, and write that into the TIFL header for the benefit of other
programs that may rely on having the correct native app name; don't
rely on the OS plus ticonv making a correct round-trip conversion.