libtifiles: add version field to backup files.
For the 83+ family, this indicates what OS versions are compatible
with the backup.  In particular, backups from OS 2.48 and earlier are
not compatible with OS 2.53 and later, so the later OSes set the
backup version to 6 or 7, and refuse to install "older" backups.
Likewise, OS 4.x on the 84+CSE uses and requires a backup version of
10 (0xa).

We continue to use the old format for versions 0 to 5, since all of
those versions are essentially compatible.  For versions 6 and later,
the backup header is extended to 12 bytes as in the DBUS protocol.