libticalcs: support sending backups to 84+ / 84+CSE / 82A / 84+T DUSB.
Sending backups via "silent" DUSB commands seems not to work for some
reason.  Instead, we use silent remote-control commands to enter
calc-to-calc linking mode, and transfer the backup "non-silently".
Normally this would also require the user to press a key on the
calculator to accept the backup; we skip this step as well by sending
another remote-control keypress.