libticalcs: correct format of 84+ DUSB remote keycode packets.
For two-byte keycodes, all OS versions require the "low" byte
(keyExtend value) in the first byte following the "action" code, and
the "high" byte (kbdKey value) in the second byte.

For one-byte keycodes, the Z80 OSes seem to accept packets with the
keycode in *either* the first or the second byte (while the other byte
is set to zero.)  However, the 84+CE requires it to be in the second
byte.