libticables: Introduce a cable for using GPIO pins on RPi SBCs
With this cable, a calculator can be connected to pins 16 (Tip/Red) and 18
(Ring/White) ideally through a ~1k resistor. Transmission happens by switching
the lines between tristate input and 0V output states.

In case the voltage of the calculator exceeds 3.3V, diodes should be used
in addition to the resistors to limit the line voltage to 3V3 to avoid
damage.